chore: move native polyfills to entrypoint
